Extracts feedController from userController
