+++ semana18/cookenu/src/routes.ts
@@ -3,12 +3,14 @@ import UserController from "./controllers/userController";
 import RecipeController from "./controllers/recipeController";
 import AuthenticationUserController from "./controllers/authenticationUserController";
 import FollowsController from "./controllers/followsController";
+import FeedController from "./controllers/feedController";
 
 const routes: Router = express.Router();
 const userController = new UserController();
 const recipeController = new RecipeController();
 const authenticationUserController = new AuthenticationUserController();
 const followsController = new FollowsController();
+const feedController = new FeedController();
 
 routes.get("/ping", (_, res) => {
   res.send({ message: "pong" });
@@ -21,7 +23,7 @@ routes.put("/reset", authenticationUserController.resetPassword);
 routes.post("/user/follow", followsController.follow);
 routes.post("/user/unfollow", followsController.unfollow);
 
-routes.get("/user/feed", userController.feed);
+routes.get("/user/feed", feedController.feed);
 
 routes.get("/user/profile", userController.getProfile);
 routes.get("/user/:id", userController.getProfileById);
